fix swapfile

+++ profiles/cloud/filesystems.nix
@@ -42,13 +42,16 @@ in
       ${pkgs.gptfdisk}/bin/sgdisk -e -d 2 -n 2:0:0 -c 2:NIXOS -p /dev/vda
       ${pkgs.util-linux}/bin/partx -u /dev/vda
       btrfs filesystem resize max /boot
-      truncate -s 0 /swap/swapfile
-      chattr +C /swap/swapfile
-      btrfs property set /swap/swapfile compression none
-      dd if=/dev/zero of=/swap/swapfile bs=1M count=1024
-      chmod 0600 /swap/swapfile
-      mkswap /swap/swapfile
+      if [ ! -f "/swap/swapfile" ]; then
+        truncate -s 0 /swap/swapfile
+        chattr +C /swap/swapfile
+        btrfs property set /swap/swapfile compression none
+        dd if=/dev/zero of=/swap/swapfile bs=1M count=1024
+        chmod 0600 /swap/swapfile
+        mkswap /swap/swapfile
+      fi
+      swapon /swap/swapfile
     '';
 
-  swapDevices = [ { device = "/swap/swapfile"; size = 1024; } ];
+  # swapDevices = [ { device = "/swap/swapfile"; size = 1024; } ];
 }
